+++ packages/api/blocks-api/src/blocks/YouTubeVideoBlock/youtube-video.block.ts
@@ -1,8 +1,9 @@
-import { IsBoolean, IsOptional, IsString, Matches } from "class-validator";
+import { IsBoolean, IsOptional, IsString } from "class-validator";
 
 import { typesafeMigrationPipe } from "../../migrations/typesafeMigrationPipe";
 import { BlockData, BlockDataInterface, BlockInput, createBlock, inputToData } from "../block";
 import { BlockField } from "../decorators/field";
+import { IsValidYoutubeIdentifier } from "../validator/is-valid-youtube-identifier";
 import { RemoveAspectRatioMigration } from "./migrations/1-remove-aspect-ratio.migration";
 
 class YouTubeVideoBlockData extends BlockData {
@@ -23,10 +24,7 @@ class YouTubeVideoBlockInput extends BlockInput {
     @IsOptional()
     @IsString()
     @BlockField({ nullable: true })
-    // regex from https://stackoverflow.com/a/51870158
-    @Matches(
-        /(https?:\/\/)?(((m|www)\.)?(youtube(-nocookie)?|youtube.googleapis)\.com.*(v\/|v=|vi=|vi\/|e\/|embed\/|user\/.*\/u\/\d+\/)|youtu\.be\/)([_0-9a-zA-Z-]+)/,
-    )
+    @IsValidYoutubeIdentifier()
     youtubeIdentifier?: string;
 
     @IsBoolean()