+++ packages/admin/admin-theme/CHANGELOG.md
@@ -1,6 +1,5 @@
 # @comet/admin-theme
 
-<<<<<<< HEAD
 ## 7.0.0-beta.3
 
 ### Patch Changes
@@ -200,7 +199,7 @@
 ### Patch Changes
 
 -   @comet/admin-icons@7.0.0-beta.0
-=======
+
 ## 6.15.1
 
 ### Patch Changes
@@ -213,7 +212,6 @@
 
 -   Updated dependencies [406027806]
     -   @comet/admin-icons@6.15.0
->>>>>>> main
 
 ## 6.14.1