+++ packages/admin/admin-theme/CHANGELOG.md
@@ -200,6 +200,19 @@
 
 -   @comet/admin-icons@7.0.0-beta.0
 
+## 6.15.1
+
+### Patch Changes
+
+-   @comet/admin-icons@6.15.1
+
+## 6.15.0
+
+### Patch Changes
+
+-   Updated dependencies [406027806]
+    -   @comet/admin-icons@6.15.0
+
 ## 6.14.1
 
 ### Patch Changes